services/worker/requirements.txt
libtorrent
redis>=5.0
python-dotenv
httpx